split builder out of parser
